requirements.txt
pynacl>=1.4.0
ipaddress>=1.0.23 